Extract config struct to config.rs
