+++ crates/storage/db/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod config;
 pub mod rocks;
 pub mod storage;
