+++ crates/tracing/src/lib.rs
@@ -1,35 +1,10 @@
+pub mod config;
+
 use std::io;
-use std::path::PathBuf;
 
+use config::TracingConfig;
 use rolling_file::{RollingConditionBasic, RollingFileAppender};
-use serde::{Deserialize, Serialize};
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt, EnvFilter, Layer};
-
-#[derive(Debug, Clone, Default, Deserialize, PartialEq, Eq, Serialize)]
-pub struct TracingConfig {
-    pub path: PathBuf,
-    pub max_size_bytes: u64,
-    pub max_files: usize,
-}
-
-impl TracingConfig {
-    pub fn new(root_path: PathBuf) -> Self {
-        let log_dir = root_path.join(Self::log_file_dir());
-        Self {
-            path: log_dir,
-            max_size_bytes: 200,
-            max_files: 5,
-        }
-    }
-
-    fn log_file_dir() -> PathBuf {
-        "logs".into()
-    }
-
-    pub fn log_file_name(&self) -> PathBuf {
-        self.path.join("ramd.log")
-    }
-}
 
 /// Init basic fmt logger that writes to console
 ///
@@ -55,10 +30,10 @@ pub fn init(config: &TracingConfig) {
             .with_writer(file_appender)
             .with_filter(EnvFilter::builder().from_env_lossy())
             .boxed(),
-        tracing_journald::layer()
-            .expect("Faiiled to init journald layer")
-            .with_filter(EnvFilter::builder().from_env_lossy())
-            .boxed(),
+        // tracing_journald::layer()
+        //     .expect("Failed to init journald layer")
+        //     .with_filter(EnvFilter::builder().from_env_lossy())
+        //     .boxed(),
     ];
 
     let _ = tracing_subscriber::registry().with(layers).try_init();
